Move Logger class to src/components/util
